requirements.txt
django>=4.2
djangorestframework>=3.14.0
psycopg2-binary>=2.9.6
dj-database-url>=2.1.0
python-dotenv>=1.0.0
requests>=2.31.0
httpx>=0.24.1
whitenoise>=6.5.0
gunicorn>=20.1.0
pytest>=7.4.0
pytest-django>=4.5.2
black>=23.7.0
flake8>=6.1.0
isort>=5.12.0
